pyproject.toml
[project]
name = "mihome-exporter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "mijiaapi>=2.0.1",
    "prometheus-client>=0.22.1",
    "pydantic>=2.11.7",
    "python-json-logger>=3.3.0",
    "pyyaml>=6.0.2",
]

[tool.uv]
index-url = "https://pypi.tuna.tsinghua.edu.cn/simple"
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/*"]
